fix: fix circular import in train.py, predict.py
